client_app/streamlit_requirements.txt
streamlit
Pillow
opencv-python
distinctipy
requests
numpy